Share `namespaceRelationships` logic between permissions and translations
Change-type: patch
